requirements.txt
playwright==1.49.1
youtube-transcript-api==0.6.3
chromadb==0.5.23
sentence-transformers==3.3.1
anthropic==0.43.0
python-dotenv==1.0.1
rich==13.9.4
